+++ website/src/App.js
@@ -1,6 +1,7 @@
 import { Home } from "@components/Home";
 import { Banner } from "@components/Banner";
 import { Dashboard } from "@admin/Dashboard";
+import { Applicants } from "@admin/Applicants";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 
 import './App.css';
@@ -11,6 +12,7 @@ function App() {
       <BrowserRouter>
         <Banner/>
         <Switch>
+          <Route path="/admin/applications"><Applicants/></Route>
           <Route path="/admin"><Dashboard/></Route>
           <Route path="/"><Home /></Route>
         </Switch>