+++ CHANGELOG.md
@@ -21,5 +21,6 @@ All notable changes to this project will be documented in this file.
 - Added mobile screen to Scan QR codes ([0f56699](https://github.com/coffeexcode/capstone/commit/0f56699db40f170264c931d6813e8c88da64d26d))
 - Added mobile screen to view the schedule of a conference ([4ffa223](https://github.com/coffeexcode/capstone/commit/4ffa2238b79ee4b5540eabbee2a511331ff47d6f))
 - Added mobile screen to view profile ([#1](https://github.com/coffeexcode/capstone/pull/1))
+- Added website page for managing event applications or registrations ([#4](https://github.com/coffeexcode/capstone/pull/4))
 
 ### Changed